Refactored datastore session classes
